# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel V12 Kanban Task Management API - FSJ28",
        "description": "API documentation for the Kanban task management system. Laravel 12.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API V1 Base URL"
        }
    ],
    "paths": {
...
        "/tasks": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Get all tasks with user names",
                "description": "Retrieves a list of all tasks, including the name of the assigned user.",
                "operationId": "6a93d54730cde509d735e538ab6800f6",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "tasks"
                ],
                "summary": "Create a new task",
                "description": "Creates a new task in the system.",
                "operationId": "9b481c0ed5fdfa1a592d70ba39a348ee",
                "requestBody": {
                    "description": "Task data for creation",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "status",
                                    "priority",
                                    "due_date",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Fix critical bug"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "The production database is failing."
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "en proceso",
                                            "completada"
                                        ],
                                        "example": "pendiente"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "baja",
                                            "media",
                                            "alta"
                                        ],
                                        "example": "alta"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-15"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Task created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/tasks/{taskId}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Get a specific task by ID",
                "description": "Returns a single task object based on the provided ID.",
                "operationId": "7359a021c9cb101106dcbcd751cf16e3",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "ID of the task to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error (e.g., Task ID not found)"
                    }
                }
            },
            "patch": {
                "tags": [
                    "tasks"
                ],
                "summary": "Update an existing task",
                "description": "Updates an existing task by ID. Requires authentication.",
                "operationId": "f5e08ba7dc8afb612806660af4fc36f7",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "ID of the task to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Fields to update (partial update supported)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Fix critical bug (completed)"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "en proceso",
                                            "completada"
                                        ],
                                        "example": "completada"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Task updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (Missing or invalid token)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/filter": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Filter tasks by status or priority",
                "description": "Returns a list of tasks filtered by optional status and/or priority query parameters.",
                "operationId": "da811e6e390a05d45d40d35df9ee4b7e",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by task status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pendiente",
                                "en proceso",
                                "completada"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filter by task priority",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "baja",
                                "media",
                                "alta"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No tasks found matching the criteria"
                    },
                    "422": {
                        "description": "Invalid filter parameter"
                    }
                }
            }
        },
        "/tasks/remaining-days/{taskId}": {
            "get": {
                "tags": [
                    "tasks"
                ],
                "summary": "Calculate remaining days until due date",
                "description": "Calculates the difference in days between today and the task's due date. Requires authentication.",
                "operationId": "9283d6fc52eafbb8cfe307aea482ef8b",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "task_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Implement Task API"
                                        },
                                        "due_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-10-30"
                                        },
                                        "remaining_days": {
                                            "description": "Days remaining (negative if overdue)",
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "Aun estas a tiempo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (Missing or invalid token)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Task": {
                "title": "Task",
                "description": "Task model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Implement Task API"
                    },
                    "description": {
                        "type": "string",
                        "example": "Develop CRUD operations for the Task resource."
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "en proceso",
                            "completada"
                        ],
                        "example": "en proceso"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "baja",
                            "media",
                            "alta"
                        ],
                        "example": "alta"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-30"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "user": {
                        "description": "Associated user name (only in index/list view)",
                        "type": "string",
                        "readOnly": "true",
                        "example": "John Doe"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the Bearer token (JWT) returned at login.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}